Revert "troff: Fix -Wsign-compare"
This reverts commit 9d3c9d451264d412cd570e1b84bd03ed9f70df6e.

+++ troff/n5.c
@@ -875,7 +875,7 @@ loop:	for (i = 0; i < NTM - 5 - mb_cur_max; ) {
 			j = ':';
 		else if (istrans(c))
 			j = ')';
-		else if (j >= 0 && j < (int)sizeof tmtab && tmtab[j])
+		else if (j >= 0 && j < sizeof tmtab && tmtab[j])
 			j = tmtab[j];
 		else if (j == ACUTE)
 			j = '\'';
@@ -2273,9 +2273,9 @@ casemc(void)
 
 
 static void
-propchar(size_t *tp)
+propchar(int *tp)
 {
-	size_t	c, *tpp;
+	int	c, *tpp;
 	tchar	i;
 
 	if (skip(0)) {
